pyproject.toml
[project]
name = "oz-externship"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "uvicorn[standard] (>=0.34.3,<0.35.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "sentry-sdk[django] (>=2.29.1,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "django-redis (>=5.4.0,<6.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "django-filter (>=25.1,<26.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "boto3 (>=1.39.2,<2.0.0)",
    "botocore (>=1.39.2,<2.0.0)",
    "twilio (>=9.6.4,<10.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "channels (>=4.2.2,<5.0.0)",
    "daphne (>=4.2.0,<5.0.0)",
    "channels-redis (>=4.2.1,<5.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
]

[tool.poetry.group.dev.dependencies]
djangorestframework-stubs = "^3.16.0"
drf-spectacular = "^0.28.0"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.16.0"
coverage = "^7.8.2"
django-debug-toolbar = "^5.2.0"
django-stubs = "^5.2.0"
boto3-stubs = "^1.39.0"
twilio-stubs = "^0.2.0"
types-python-dateutil = "^2.9.0.20250708"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile="black"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
python_version = 3.12
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "config.settings.local"

[tool.coverage.run]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/test_*.py",
    "*/settings/*",
    "*/manage.py",
    "*/urls.py",
    "*/wsgi.py",
    "*/asgi.py",
]

[tool.coverage.report]
fail_under = 80
show_missing = true
